+++ sdk/cognitivelanguage/azure-ai-language-conversations/azure/ai/language/conversations/_operations/_operations.py
@@ -231,6 +231,7 @@ class ConversationAnalysisClientOperationsMixin(ConversationAnalysisClientMixinA
 
         Example:
             .. code-block:: python
+
                 # The response is polymorphic. The following are possible polymorphic responses based
                   off discriminator "kind":
 
@@ -309,6 +310,7 @@ class ConversationAnalysisClientOperationsMixin(ConversationAnalysisClientMixinA
 
         Example:
             .. code-block:: python
+
                 # The response is polymorphic. The following are possible polymorphic responses based
                   off discriminator "kind":