+++ sdk/ml/azure-ai-ml/azure/ai/ml/entities/_component/component.py
@@ -1,7 +1,6 @@
 # ---------------------------------------------------------
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # ---------------------------------------------------------
-import os
 import tempfile
 from contextlib import contextmanager
 from os import PathLike
@@ -32,9 +31,10 @@ from azure.ai.ml.entities._assets.asset import Asset
 from azure.ai.ml.entities._inputs_outputs import Input, Output
 from azure.ai.ml.entities._mixins import RestTranslatableMixin, TelemetryMixin, YamlTranslatableMixin
 from azure.ai.ml.entities._system_data import SystemData
-from azure.ai.ml.entities._util import find_type_in_override, _copy_folder_ignore_pycache
+from azure.ai.ml.entities._util import find_type_in_override
 from azure.ai.ml.entities._validation import SchemaValidatableMixin, MutableValidationResult
 from azure.ai.ml.exceptions import ErrorCategory, ErrorTarget, ValidationException
+from .code import ComponentIgnoreFile
 
 # pylint: disable=protected-access, redefined-builtin
 # disable redefined-builtin to use id/type as argument name
@@ -468,6 +468,9 @@ class Component(
             component["type"] = NodeType.COMMAND
             component["inputs"] = component.pop("source")
             component["outputs"] = dict({"output": component.pop("output")})
+            # method _to_dict() will remove empty keys
+            if "tags" not in component:
+                component["tags"] = {}
             component["tags"]["component_type_overwrite"] = NodeType.IMPORT
             component["command"] = NodeType.IMPORT
 
@@ -509,37 +512,18 @@ class Component(
         if not hasattr(self, "code"):
             raise ValueError(f"{self.__class__} does not have attribute code.")
         code = getattr(self, "code")
-        # special check for git path code value
-        if code is not None and isinstance(code, str) and code.startswith("git+"):
-            yield Code(base_path=self._base_path, path=code)
-        elif code is not None and os.path.isfile(code):
-            yield Code(base_path=self._base_path, path=code)
-        else:
+        if code is None:
+            # Hack: when code not specified, we generated a file which contains
+            # COMPONENT_PLACEHOLDER as code
+            # This hack was introduced because job does not allow running component without a
+            # code, and we need to make sure when component updated some field(eg: description),
+            # the code remains the same. Benefit of using a constant code for all components
+            # without code is this will generate same code for anonymous components which
+            # enables component reuse
             with tempfile.TemporaryDirectory() as tmp_dir:
-                if code is None:
-                    # Hack: when code not specified, we generated a file which contains
-                    # COMPONENT_PLACEHOLDER as code
-                    # This hack was introduced because job does not allow running component without a
-                    # code, and we need to make sure when component updated some field(eg: description),
-                    # the code remains the same. Benefit of using a constant code for all components
-                    # without code is this will generate same code for anonymous components which
-                    # enables component reuse
-                    code = Path(tmp_dir) / COMPONENT_PLACEHOLDER
-                    with open(code, "w") as f:
-                        f.write(COMPONENT_CODE_PLACEHOLDER)
-                    yield Code(base_path=self._base_path, path=code)
-                else:
-                    # copy to temp folder to filter potential __pycache__
-                    # note that the dst here is one level deeper than the temp folder,
-                    # when upload code asset, the URL contains one level higher than the file,
-                    # if we simply copy to temp folder, the URL will contain random temp folder name,
-                    # that might result in unexpected issues and also break related tests.
-                    # therefore we copy deeper to avoid this break.
-                    src_path = Path(self._base_path) / code
-                    # .name will return empty string for UNC drive names
-                    # so we need src_path.resolve() here to avoid empty string
-                    # that leads to FileExistsError during shutil.copytree
-                    # TODO(2056980): replace temp code folder name with constant value
-                    dst_path = Path(tmp_dir) / src_path.resolve().name
-                    _copy_folder_ignore_pycache(src_path, dst_path)
-                    yield Code(base_path=self._base_path, path=dst_path)
+                code = Path(tmp_dir) / COMPONENT_PLACEHOLDER
+                with open(code, "w") as f:
+                    f.write(COMPONENT_CODE_PLACEHOLDER)
+                yield Code(base_path=self._base_path, path=code)
+        else:
+            yield Code(base_path=self._base_path, path=code, ignore_file=ComponentIgnoreFile(code))