+++ sdk/ml/azure-ai-ml/tests/dsl/e2etests/test_dsl_pipeline.py
@@ -32,6 +32,7 @@ from devtools_testutils import AzureRecordedTestCase
 from pipeline_job.e2etests.test_pipeline_job import assert_job_input_output_types
 from test_utilities.utils import _PYTEST_TIMEOUT_METHOD, omit_with_wildcard, assert_job_cancel, sleep_if_live
 
+from devtools_testutils import is_live
 from .._util import _DSL_TIMEOUT_SECOND
 
 tests_root_dir = Path(__file__).parent.parent.parent
@@ -200,10 +201,6 @@ class TestDSLPipeline(AzureRecordedTestCase):
         omit_fields = ["componentId", "_source", "properties"]
         component_job_dict = pydash.omit(component_job_dict, *omit_fields)
         assert component_job_dict == {
-            "computeId": None,
-            "display_name": None,
-            "distribution": None,
-            "environment_variables": {},
             "inputs": {
                 "component_in_number": {"job_input_type": "literal", "value": "10"},
                 "component_in_path": {
@@ -211,15 +208,10 @@ class TestDSLPipeline(AzureRecordedTestCase):
                     "uri": "https://dprepdata.blob.core.windows.net/demo/Titanic.csv",
                 },
             },
-            "limits": None,
-            "name": None,
-            "outputs": {},
-            "resources": {"instance_count": 1, "properties": {}},
-            "tags": {},
+            "resources": {"instance_count": 1},
             "type": "command",
         }
 
-    @pytest.mark.skip(reason="migration skip: gpu-cluster is not available yet.")
     def test_distribution_components(self, client: MLClient, randstr: Callable[[str], str]) -> None:
         mpi_func = load_component(source=str(components_dir / "helloworld_component_mpi.yml"))
         pytorch_func = load_component(source=str(components_dir / "helloworld_component_pytorch.yml"))
@@ -437,10 +429,6 @@ class TestDSLPipeline(AzureRecordedTestCase):
             "jobs": {
                 "node1": {
                     "type": "command",
-                    "computeId": None,
-                    "display_name": None,
-                    "distribution": None,
-                    "environment_variables": {},
                     "inputs": {
                         "component_in_number": {
                             "job_input_type": "literal",
@@ -448,18 +436,11 @@ class TestDSLPipeline(AzureRecordedTestCase):
                         },
                         "component_in_path": {"job_input_type": "literal", "value": "${{parent.inputs.job_in_path}}"},
                     },
-                    "limits": None,
                     "name": "node1",
-                    "outputs": {},
-                    "resources": None,
-                    "tags": {},
                 },
                 "node2": {
                     "type": "command",
-                    "computeId": None,
-                    "display_name": None,
                     "distribution": {"distribution_type": "PyTorch", "process_count_per_instance": 2},
-                    "environment_variables": {},
                     "inputs": {
                         "component_in_number": {"job_input_type": "literal", "value": "2"},
                         "component_in_path": {
@@ -467,14 +448,10 @@ class TestDSLPipeline(AzureRecordedTestCase):
                             "value": "${{parent.jobs.node1.outputs.component_out_path}}",
                         },
                     },
-                    "limits": None,
                     "name": "node2",
-                    "outputs": {},
-                    "resources": {"instance_count": 2, "properties": {}},
-                    "tags": {},
+                    "resources": {"instance_count": 2},
                 },
                 "node3": {
-                    "computeId": None,
                     "display_name": "command-function-job",
                     "distribution": {"distribution_type": "PyTorch", "process_count_per_instance": 2},
                     "environment_variables": {"environ": "val"},
@@ -485,13 +462,11 @@ class TestDSLPipeline(AzureRecordedTestCase):
                             "value": "${{parent.jobs.node2.outputs.component_out_path}}",
                         },
                     },
-                    "limits": None,
                     "name": "node3",
                     "outputs": {
                         "component_out_path": {"type": "literal", "value": "${{parent.outputs.pipeline_job_out}}"}
                     },
-                    "resources": {"instance_count": 2, "properties": {}},
-                    "tags": {},
+                    "resources": {"instance_count": 2},
                     "type": "command",
                 },
             },
@@ -574,22 +549,13 @@ class TestDSLPipeline(AzureRecordedTestCase):
             },
             "jobs": {
                 "node1": {
-                    "computeId": None,
-                    "display_name": None,
-                    "distribution": None,
-                    "environment_variables": {},
                     "inputs": {
                         "component_in_path": {"job_input_type": "literal", "value": "${{parent.inputs.job_in_file}}"}
                     },
-                    "limits": None,
                     "name": "node1",
-                    "outputs": {},
-                    "resources": None,
-                    "tags": {},
                     "type": "command",
                 },
                 "node2": {
-                    "computeId": None,
                     "display_name": "command_with_optional_inputs",
                     "distribution": {"distribution_type": "PyTorch", "process_count_per_instance": 2},
                     "environment_variables": {"environ": "val"},
@@ -599,13 +565,11 @@ class TestDSLPipeline(AzureRecordedTestCase):
                             "value": "${{parent.jobs.node1.outputs.component_out_path}}",
                         }
                     },
-                    "limits": None,
                     "name": "node2",
                     "outputs": {
                         "component_out_path": {"type": "literal", "value": "${{parent.outputs.pipeline_output}}"}
                     },
-                    "resources": {"instance_count": 2, "properties": {}},
-                    "tags": {},
+                    "resources": {"instance_count": 2},
                     "type": "command",
                 },
             },
@@ -614,21 +578,18 @@ class TestDSLPipeline(AzureRecordedTestCase):
         }
         assert expected_job == actual_job
 
-    @pytest.mark.skip("Skip for compute resource not ready.")
     def test_spark_with_optional_inputs(self, randstr: Callable[[str], str], client: MLClient):
         component_yaml = "./tests/test_configs/dsl_pipeline/spark_job_in_pipeline/component_with_optional_inputs.yml"
         spark_with_optional_inputs_component_func = load_component(source=component_yaml)
-        synapse_compute_name = "spark31"
 
         @dsl.pipeline(
             name=f"test_optional_input_component_pipeline_" + randstr("pipeline_name"),
             description="The spark node with optional inputs",
             tags={"owner": "sdkteam", "tag": "tagvalue"},
-            compute=synapse_compute_name,
         )
         def sample_pipeline(job_in_file, sample_rate):
             node1 = spark_with_optional_inputs_component_func(input1=job_in_file, sample_rate=sample_rate)
-            node1.compute = synapse_compute_name
+            node1.resources = {"instance_type": "standard_e4s_v3", "runtime_version": "3.1.0"}
             return {"pipeline_output": node1.outputs.output1}
 
         pipeline = sample_pipeline(
@@ -646,7 +607,6 @@ class TestDSLPipeline(AzureRecordedTestCase):
         expected_job = {
             "description": "The spark node with optional inputs",
             "tags": {"owner": "sdkteam", "tag": "tagvalue"},
-            "compute_id": "spark31",
             "is_archived": False,
             "job_type": "Pipeline",
             "inputs": {
@@ -655,12 +615,10 @@ class TestDSLPipeline(AzureRecordedTestCase):
             },
             "jobs": {
                 "node1": {
-                    "archives": None,
                     "args": "--input1 ${{inputs.input1}} --output2 "
                     "${{outputs.output1}} --my_sample_rate "
                     "${{inputs.sample_rate}} $[[--input_optional "
                     "${{inputs.input_optional}}]]",
-                    "computeId": "spark31",
                     "conf": {
                         "spark.driver.cores": 1,
                         "spark.driver.memory": "2g",
@@ -671,23 +629,19 @@ class TestDSLPipeline(AzureRecordedTestCase):
                         "spark.executor.instances": 1,
                         "spark.executor.memory": "2g",
                     },
-                    "display_name": None,
                     "entry": {
                         "file": "sampleword_with_optional_input.py",
                         "spark_job_entry_type": "SparkJobPythonEntry",
                     },
-                    "files": None,
-                    "identity": {"identity_type": "Managed"},
+                    "identity": {"identity_type": "UserIdentity"},
                     "inputs": {
                         "input1": {"job_input_type": "literal", "value": "${{parent.inputs.job_in_file}}"},
                         "sample_rate": {"job_input_type": "literal", "value": "${{parent.inputs.sample_rate}}"},
                     },
-                    "jars": None,
                     "name": "node1",
                     "outputs": {"output1": {"type": "literal", "value": "${{parent.outputs.pipeline_output}}"}},
-                    "py_files": None,
-                    "resources": None,
-                    "tags": {},
+                    'resources': {'instance_type': 'standard_e4s_v3',
+                                  'runtime_version': '3.1.0'},
                     "type": "spark",
                 }
             },
@@ -740,10 +694,6 @@ class TestDSLPipeline(AzureRecordedTestCase):
             "jobs": {
                 "node1": {
                     "type": "command",
-                    "computeId": None,
-                    "display_name": None,
-                    "distribution": None,
-                    "environment_variables": {},
                     "inputs": {
                         "component_in_number": {
                             "job_input_type": "literal",
@@ -751,19 +701,11 @@ class TestDSLPipeline(AzureRecordedTestCase):
                         },
                         "component_in_path": {"job_input_type": "literal", "value": "${{parent.inputs.job_in_path}}"},
                     },
-                    "limits": None,
                     "name": "node1",
                     "new_field": "val",
-                    "outputs": {},
-                    "resources": None,
-                    "tags": {},
                 },
                 "node2": {
                     "type": "command",
-                    "computeId": None,
-                    "display_name": None,
-                    "distribution": None,
-                    "environment_variables": {},
                     "inputs": {
                         "component_in_number": {
                             "job_input_type": "literal",
@@ -771,11 +713,7 @@ class TestDSLPipeline(AzureRecordedTestCase):
                         },
                         "component_in_path": {"job_input_type": "literal", "value": "${{parent.inputs.job_in_path}}"},
                     },
-                    "limits": None,
                     "name": "node2",
-                    "outputs": {},
-                    "resources": None,
-                    "tags": {},
                 },
             },
             "outputs": {},
@@ -1352,6 +1290,8 @@ class TestDSLPipeline(AzureRecordedTestCase):
         # assert "optional_param_with_default" not in pipeline_job.inputs
         # assert "optional_param_with_default" not in next(pipeline_job.jobs.values().__iter__()).inputs
 
+    @pytest.mark.disable_mock_code_hash
+    @pytest.mark.skipif(condition=not is_live(), reason="reuse test, target to verify service-side behavior")
     def test_component_reuse(self, client: MLClient) -> None:
         path = "./tests/test_configs/components/helloworld_component.yml"
         component_func1 = load_component(source=path)
@@ -1379,7 +1319,8 @@ class TestDSLPipeline(AzureRecordedTestCase):
 
         assert len(component_ids) == 1, f"Got multiple component id: {component_ids} for same anon component."
 
-    @pytest.mark.skip(reason="migration skip: need to discuss reuse in current test mechanism.")
+    @pytest.mark.disable_mock_code_hash
+    @pytest.mark.skipif(condition=not is_live(), reason="reuse test, target to verify service-side behavior")
     def test_pipeline_reuse(self, client: MLClient, randstr: Callable[[str], str], randint: Callable) -> None:
         component_yaml = components_dir / "helloworld_component.yml"
         component_func1 = load_component(source=component_yaml, params_override=[{"name": randstr("component_name")}])
@@ -1494,7 +1435,6 @@ class TestDSLPipeline(AzureRecordedTestCase):
             _ = client.jobs.create_or_update(dsl_pipeline)
             mock_logging.assert_called_with("Warnings: [jobs.node1.jeff_special_option: Unknown field.]")
 
-    @pytest.mark.skip(reason="migration skip: gpu-cluster is not available yet.")
     def test_anon_component_in_pipeline(
         self, client: MLClient, randstr: Callable[[str], str], hello_world_component: Component
     ) -> None:
@@ -1600,7 +1540,7 @@ class TestDSLPipeline(AzureRecordedTestCase):
 
         # check required fields in job dict
         job_dict = pipeline_job._to_dict()
-        expected_keys = ["status", "properties", "tags", "creation_context"]
+        expected_keys = ["status", "properties", "creation_context"]
         for k in expected_keys:
             assert k in job_dict.keys(), f"failed to get {k} in {job_dict}"
 
@@ -1630,7 +1570,7 @@ class TestDSLPipeline(AzureRecordedTestCase):
         pipeline_job = assert_job_cancel(pipeline, client, experiment_name="parallel_in_pipeline")
         # check required fields in job dict
         job_dict = pipeline_job._to_dict()
-        expected_keys = ["status", "properties", "tags", "creation_context"]
+        expected_keys = ["status", "properties", "creation_context"]
         for k in expected_keys:
             assert k in job_dict.keys(), f"failed to get {k} in {job_dict}"
 
@@ -1711,7 +1651,6 @@ class TestDSLPipeline(AzureRecordedTestCase):
             },
             "jobs": {
                 "node1": {
-                    "computeId": None,
                     "input_data": "${{inputs.job_data_path}}",
                     "display_name": "my-evaluate-job",
                     "inputs": {
@@ -1722,14 +1661,12 @@ class TestDSLPipeline(AzureRecordedTestCase):
                     },
                     "name": "node1",
                     "mini_batch_size": 5,
-                    "partition_keys": None,
-                    "retry_settings": None,
                     "logging_level": "DEBUG",
                     "max_concurrency_per_instance": 1,
                     "error_threshold": 1,
                     "mini_batch_error_threshold": 1,
                     "outputs": {"job_output_path": {"type": "literal", "value": "${{parent.outputs.pipeline_output}}"}},
-                    "resources": {"instance_count": 2, "properties": {}},
+                    "resources": {"instance_count": 2},
                     "type": "parallel",
                     "task": {
                         "type": "run_function",
@@ -1738,8 +1675,6 @@ class TestDSLPipeline(AzureRecordedTestCase):
                         "program_arguments": "--job_output_path ${{outputs.job_output_path}}",
                         "environment": "azureml:AzureML-sklearn-0.24-ubuntu18.04-py37-cpu:5",
                     },
-                    "tags": {},
-                    "environment_variables": {},
                 },
             },
             "outputs": {
@@ -1754,7 +1689,7 @@ class TestDSLPipeline(AzureRecordedTestCase):
 
         # check required fields in job dict
         job_dict = pipeline_job._to_dict()
-        expected_keys = ["status", "properties", "tags", "creation_context"]
+        expected_keys = ["status", "properties", "creation_context"]
         for k in expected_keys:
             assert k in job_dict.keys(), f"failed to get {k} in {job_dict}"
 
@@ -1840,9 +1775,7 @@ class TestDSLPipeline(AzureRecordedTestCase):
             },
             "jobs": {
                 "parallel_node": {
-                    "computeId": None,
                     "input_data": "${{inputs.job_data_path}}",
-                    "display_name": None,
                     "inputs": {
                         "job_data_path": {
                             "job_input_type": "literal",
@@ -1850,24 +1783,19 @@ class TestDSLPipeline(AzureRecordedTestCase):
                         }
                     },
                     "mini_batch_size": 5,
-                    "retry_settings": None,
-                    "logging_level": None,
                     "max_concurrency_per_instance": 1,
-                    "error_threshold": None,
                     "mini_batch_error_threshold": 1,
                     "name": "parallel_node",
                     "outputs": {
                         "job_output_path": {"value": "${{parent.outputs.pipeline_job_out}}", "type": "literal"}
                     },
-                    "resources": {"instance_count": 2, "properties": {}},
+                    "resources": {"instance_count": 2},
                     "type": "parallel",
                     "task": {
                         "type": "run_function",
                         "entry_script": "score.py",
                         "program_arguments": "--job_output_path ${{outputs.job_output_path}}",
                     },
-                    "tags": {},
-                    "environment_variables": {},
                 },
             },
             "outputs": {
@@ -1925,35 +1853,22 @@ class TestDSLPipeline(AzureRecordedTestCase):
                 "batch_inference_node1": {
                     "type": "parallel",
                     "name": "batch_inference_node1",
-                    "display_name": None,
-                    "tags": {},
-                    "computeId": None,
                     "inputs": {
                         "job_data_path": {"job_input_type": "literal", "value": "${{parent.inputs.job_data_path}}"}
                     },
-                    "outputs": {},
                     "mini_batch_size": 1,
-                    "partition_keys": None,
                     "task": {
                         "type": "run_function",
                         "entry_script": "score.py",
                         "program_arguments": "--job_output_path ${{outputs.job_output_path}}",
                     },
                     "input_data": "${{inputs.job_data_path}}",
-                    "retry_settings": None,
-                    "logging_level": None,
-                    "resources": {"instance_count": 2, "properties": {}},
+                    "resources": {"instance_count": 2},
                     "max_concurrency_per_instance": 1,
-                    "error_threshold": None,
                     "mini_batch_error_threshold": 1,
-                    "environment_variables": {},
                 },
                 "convert_data_node": {
-                    "environment_variables": {},
                     "name": "convert_data_node",
-                    "display_name": None,
-                    "tags": {},
-                    "computeId": None,
                     "inputs": {
                         "input_data": {
                             "job_input_type": "literal",
@@ -1961,17 +1876,11 @@ class TestDSLPipeline(AzureRecordedTestCase):
                         }
                     },
                     "outputs": {"file_output_data": {"job_output_type": "mltable"}},
-                    "resources": None,
-                    "distribution": None,
-                    "limits": None,
                     "type": "command",
                 },
                 "batch_inference_node2": {
                     "type": "parallel",
                     "name": "batch_inference_node2",
-                    "display_name": None,
-                    "tags": {},
-                    "computeId": None,
                     "inputs": {
                         "job_data_path": {
                             "job_input_type": "literal",
@@ -1981,20 +1890,15 @@ class TestDSLPipeline(AzureRecordedTestCase):
                     },
                     "outputs": {"job_output_path": {"value": "${{parent.outputs.job_out_data}}", "type": "literal"}},
                     "mini_batch_size": 1,
-                    "partition_keys": None,
                     "task": {
                         "type": "run_function",
                         "entry_script": "score.py",
                         "program_arguments": "--job_output_path ${{outputs.job_output_path}}",
                     },
                     "input_data": "${{inputs.job_data_path}}",
-                    "retry_settings": None,
-                    "logging_level": None,
-                    "resources": {"instance_count": 2, "properties": {}},
+                    "resources": {"instance_count": 2},
                     "max_concurrency_per_instance": 1,
-                    "error_threshold": None,
                     "mini_batch_error_threshold": 1,
-                    "environment_variables": {},
                 },
             },
             "outputs": {"job_out_data": {"mode": "Upload", "job_output_type": "uri_folder"}},
@@ -2053,14 +1957,7 @@ class TestDSLPipeline(AzureRecordedTestCase):
             "jobs": {
                 "train_with_sample_data": {
                     "type": "command",
-                    "resources": None,
-                    "distribution": None,
-                    "limits": None,
-                    "environment_variables": {},
                     "name": "train_with_sample_data",
-                    "display_name": None,
-                    "tags": {},
-                    "computeId": None,
                     "inputs": {
                         "training_data": {"job_input_type": "literal", "value": "${{parent.inputs.training_input}}"},
                         "max_epochs": {"job_input_type": "literal", "value": "${{parent.inputs.training_max_epochs}}"},
@@ -2105,14 +2002,7 @@ class TestDSLPipeline(AzureRecordedTestCase):
             "jobs": {
                 "train_with_sample_data": {
                     "type": "command",
-                    "resources": None,
-                    "distribution": None,
-                    "limits": None,
-                    "environment_variables": {},
                     "name": "train_with_sample_data",
-                    "display_name": None,
-                    "tags": {},
-                    "computeId": None,
                     "inputs": {
                         "training_data": {"job_input_type": "literal", "value": "${{parent.inputs.training_input}}"},
                         "max_epochs": {"job_input_type": "literal", "value": "${{parent.inputs.training_max_epochs}}"},
@@ -2158,14 +2048,7 @@ class TestDSLPipeline(AzureRecordedTestCase):
             "jobs": {
                 "train_with_sample_data": {
                     "type": "command",
-                    "resources": None,
-                    "distribution": None,
-                    "limits": None,
-                    "environment_variables": {},
                     "name": "train_with_sample_data",
-                    "display_name": None,
-                    "tags": {},
-                    "computeId": None,
                     "inputs": {
                         "training_data": {
                             "mode": "ReadOnlyMount",
@@ -2220,14 +2103,7 @@ class TestDSLPipeline(AzureRecordedTestCase):
             "jobs": {
                 "train_with_sample_data": {
                     "type": "command",
-                    "resources": None,
-                    "distribution": None,
-                    "limits": None,
-                    "environment_variables": {},
                     "name": "train_with_sample_data",
-                    "display_name": None,
-                    "tags": {},
-                    "computeId": None,
                     "inputs": {
                         "training_data": {
                             "mode": "ReadOnlyMount",
@@ -2283,14 +2159,8 @@ class TestDSLPipeline(AzureRecordedTestCase):
             "jobs": {
                 "train_with_sample_data": {
                     "type": "command",
-                    "resources": None,
                     "distribution": {"distribution_type": "PyTorch", "process_count_per_instance": 2},
-                    "limits": None,
-                    "environment_variables": {},
                     "name": "train_with_sample_data",
-                    "display_name": None,
-                    "tags": {},
-                    "computeId": None,
                     "inputs": {
                         "training_data": {
                             "mode": "ReadOnlyMount",
@@ -2321,20 +2191,18 @@ class TestDSLPipeline(AzureRecordedTestCase):
         }
         assert expected_job == actual_job
 
-    @pytest.mark.skip("Skip for compute resource not ready.")
     def test_spark_components(self, client: MLClient, randstr: Callable[[str], str]) -> None:
         components_dir = tests_root_dir / "test_configs/dsl_pipeline/spark_job_in_pipeline"
-        synapse_compute_name = "spark31"
         add_greeting_column = load_component(str(components_dir / "add_greeting_column_component.yml"))
         count_by_row = load_component(str(components_dir / "count_by_row_component.yml"))
 
         # Construct pipeline
-        @dsl.pipeline(compute=synapse_compute_name)
+        @dsl.pipeline()
         def spark_pipeline_from_yaml(iris_data):
             add_greeting_column_node = add_greeting_column(file_input=iris_data)
-            add_greeting_column_node.compute = synapse_compute_name
+            add_greeting_column_node.resources = {"instance_type": "standard_e4s_v3", "runtime_version": "3.1.0"}
             count_by_row_node = count_by_row(file_input=iris_data)
-            count_by_row_node.compute = synapse_compute_name
+            count_by_row_node.resources = {"instance_type": "standard_e4s_v3", "runtime_version": "3.1.0"}
             return {"output": count_by_row_node.outputs.output}
 
         pipeline = spark_pipeline_from_yaml(
@@ -2351,13 +2219,12 @@ class TestDSLPipeline(AzureRecordedTestCase):
         pipeline_job = assert_job_cancel(pipeline, client, experiment_name="spark_in_pipeline")
         # check required fields in job dict
         job_dict = pipeline_job._to_dict()
-        expected_keys = ["status", "properties", "tags", "creation_context"]
+        expected_keys = ["status", "properties", "creation_context"]
         for k in expected_keys:
             assert k in job_dict.keys(), f"failed to get {k} in {job_dict}"
 
         # original job did not change
         assert_job_input_output_types(pipeline_job)
-        assert pipeline_job.compute == synapse_compute_name
 
     def test_pipeline_with_group(self, client: MLClient):
         from enum import Enum
@@ -2465,52 +2332,25 @@ class TestDSLPipeline(AzureRecordedTestCase):
         ]
         actual_dict = omit_with_wildcard(pipeline_job._to_rest_object().as_dict()["properties"], *omit_fields)
         assert actual_dict["jobs"] == {
-            'node1': {'computeId': None,
-                      'display_name': None,
-                      'distribution': None,
-                      'environment_variables': {},
-                      'identity': {'type': 'aml_token'},
+            'node1': {'identity': {'type': 'aml_token'},
                       'inputs': {'component_in_number': {'job_input_type': 'literal',
                                                          'value': '1'},
                                  'component_in_path': {'job_input_type': 'literal',
                                                        'value': '${{parent.inputs.component_in_path}}'}},
-                      'limits': None,
                       'name': 'node1',
-                      'outputs': {},
-                      'properties': {},
-                      'resources': None,
-                      'tags': {},
                       'type': 'command'},
-            'node2': {'computeId': None,
-                      'display_name': None,
-                      'distribution': None,
-                      'environment_variables': {},
-                      'identity': {'type': 'user_identity'},
+            'node2': {'identity': {'type': 'user_identity'},
                       'inputs': {'component_in_number': {'job_input_type': 'literal',
                                                          'value': '1'},
                                  'component_in_path': {'job_input_type': 'literal',
                                                        'value': '${{parent.inputs.component_in_path}}'}},
-                      'limits': None,
                       'name': 'node2',
-                      'outputs': {},
-                      'properties': {},
-                      'resources': None,
-                      'tags': {},
                       'type': 'command'},
-            'node3': {'computeId': None,
-                      'display_name': None,
-                      'distribution': None,
-                      'environment_variables': {},
-                      'identity': {'type': 'managed_identity'},
+            'node3': {'identity': {'type': 'managed_identity'},
                       'inputs': {'component_in_number': {'job_input_type': 'literal',
                                                          'value': '1'},
                                  'component_in_path': {'job_input_type': 'literal',
                                                        'value': '${{parent.inputs.component_in_path}}'}},
-                      'limits': None,
                       'name': 'node3',
-                      'outputs': {},
-                      'properties': {},
-                      'resources': None,
-                      'tags': {},
                       'type': 'command'}
         }