+++ sdk/textanalytics/azure-ai-textanalytics/tests/test_analyze_async.py
@@ -868,7 +868,6 @@ class TestAnalyzeAsync(TextAnalyticsTest):
         assert isinstance(action_results[1][1], RecognizePiiEntitiesResult)
         assert action_results[1][1].id == "2"
 
-    @pytest.mark.skip("test is hanging with PPE")
     @TextAnalyticsPreparer()
     @TextAnalyticsClientPreparer()
     @recorded_by_proxy_async
@@ -1912,6 +1911,7 @@ class TestAnalyzeAsync(TextAnalyticsTest):
                 await poller.cancel()
             assert"Cancellation not supported by API versions v3.0, v3.1." in str(e.value)
 
+    @pytest.mark.skipif(not is_public_cloud(), reason='Usgov and China Cloud are not supported')
     @TextAnalyticsPreparer()
     @TextAnalyticsClientPreparer()
     @recorded_by_proxy_async
@@ -1961,7 +1961,7 @@ class TestAnalyzeAsync(TextAnalyticsTest):
                         assert sentence.length is not None
                     assert result.id is not None
 
-    @pytest.mark.skip("https://dev.azure.com/msazure/Cognitive%20Services/_workitems/edit/15772270")
+    @pytest.mark.skipif(not is_public_cloud(), reason='Usgov and China Cloud are not supported')
     @TextAnalyticsPreparer()
     @TextAnalyticsClientPreparer()
     @recorded_by_proxy_async
@@ -2011,6 +2011,7 @@ class TestAnalyzeAsync(TextAnalyticsTest):
                         assert sentence.length is not None
                     assert result.id is not None
 
+    @pytest.mark.skipif(not is_public_cloud(), reason='Usgov and China Cloud are not supported')
     @TextAnalyticsPreparer()
     @TextAnalyticsClientPreparer()
     @recorded_by_proxy_async
@@ -2065,6 +2066,7 @@ class TestAnalyzeAsync(TextAnalyticsTest):
                     if res.resolution_kind == "AgeResolution":
                         assert res.value == 1
 
+    @pytest.mark.skipif(not is_public_cloud(), reason='Usgov and China Cloud are not supported')
     @TextAnalyticsPreparer()
     @TextAnalyticsClientPreparer()
     @recorded_by_proxy_async
@@ -2104,6 +2106,7 @@ class TestAnalyzeAsync(TextAnalyticsTest):
                         assert context.length is not None
                     assert summary.text
 
+    @pytest.mark.skipif(not is_public_cloud(), reason='Usgov and China Cloud are not supported')
     @TextAnalyticsPreparer()
     @TextAnalyticsClientPreparer()
     @recorded_by_proxy_async
@@ -2181,13 +2184,12 @@ class TestAnalyzeAsync(TextAnalyticsTest):
                     else:
                         assert doc.detected_language.iso6391_name == "es"
 
-    @pytest.mark.skip("https://dev.azure.com/msazure/Cognitive%20Services/_workitems/edit/15816856")
+    @pytest.mark.skipif(not is_public_cloud(), reason='Usgov and China Cloud are not supported')
     @TextAnalyticsPreparer()
     @TextAnalyticsClientPreparer()
     @recorded_by_proxy_async
-    async def test_autodetect_with_default_and_script(self, client):
-        single_doc = "Tumhara naam kya hai?"
-        docs = [{"id": "1", "text": single_doc}]
+    async def test_autodetect_with_default(self, client):
+        docs = ["hello world"]
         actions=[
             RecognizeEntitiesAction(),
             ExtractKeyPhrasesAction(),
@@ -2199,12 +2201,11 @@ class TestAnalyzeAsync(TextAnalyticsTest):
             docs,
             actions,
             language="auto",
-            autodetect_default_language="en",
+            autodetect_default_language="es",
             polling_interval=self._interval(),
         )
 
         result = await poller.result()
         async for res in result:
             for doc in res:
-                assert doc.detected_language.iso6391_name == "hi"
-                assert doc.detected_language.script == "Latin"
+                assert doc.detected_language.iso6391_name == "en"